Clean up, move Anoma into provider
