+++ apps/extension/src/provider/index.ts
@@ -1,1 +1,2 @@
+export * from "./Anoma";
 export * from "./InjectedAnoma";
